Workaround declaring strnlen two times by commenting out
The fuzz target build fails because the strnlen function is declared in
system libaries and in src/compat.h. The Declaration in src/compat.h is
intended as a fallback if no other declaration is found and should not
be active. Because of some reasons it is active.
The workaround is to comment out the declaration.